code review

+++ source/MRCuda/MRCudaFastWindingNumber.cpp
@@ -89,7 +89,7 @@ void FastWindingNumber::calcFromGrid( std::vector<float>& res, const Vector3i& d
     };
     
     const Matrix4 cudaGridToMeshXf = ( gridToMeshXf == AffineXf3f{} ) ? Matrix4{} : getCudaMatrix( gridToMeshXf );
-    const size_t size = dims.x * dims.y * dims.z;
+    const size_t size = size_t( dims.x ) * dims.y * dims.z;
     data_->cudaResult.resize( size );    
     fastWindingNumberFromGridKernel( int3{ dims.x, dims.y, dims.z }, float3{ minCoord.x, minCoord.y, minCoord.z }, float3{ voxelSize.x, voxelSize.y, voxelSize.z }, cudaGridToMeshXf,
                                      data_->dipoles.data(), data_->cudaNodes.data(), data_->cudaMeshPoints.data(), data_->cudaFaces.data(), data_->cudaResult.data(), beta );
